a (+5 squashed commit)
Squashed commit:

[4fbf131] AnalizadorLéxico y Sintáctico - beta

[f8944ee] Arreglar Switch

[2981c9d] Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch angel
Your branch is ahead of 'origin/angel' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   ElecatPlusIDE/gramatica/SimpleParser.g4

[aca27bf] Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch angel
Your branch is up to date with 'origin/angel'.

Changes to be committed:
	modified:   ElecatPlusIDE/gramatica/SimpleLexer.g4
	modified:   ElecatPlusIDE/gramatica/SimpleParser.g4
	modified:   ElecatPlusIDE/src/Prueba.ecp

[e29996b] Sintaxis PreFinal

Co-authored-by: GadyelMtz <GadyelMtz@users.noreply.github.com>